NewsPaper/requirements.txt
flake8
black
django
django-allauth
django-apscheduler
django-filter
celery
django-crispy-forms
crispy_bootstrap5
django-debug-toolbar
django-modeltranslation